# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ATV API",
        "description": "API documentation for ATV application with authentication endpoints",
        "contact": {
            "email": "admin@atv.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/atvs/{id}/images": {
            "get": {
                "tags": [
                    "ATVs"
                ],
                "summary": "Get all images for a specific ATV",
                "operationId": "245e60242a746b90a368ecaef4ded3ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ATV ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by type (image, video)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "image"
                        }
                    },
                    {
                        "name": "active_only",
                        "in": "query",
                        "description": "Show only active images",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ATV images retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AtvImage"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ATV not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ATV not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/atvs/{atv_id}/images": {
            "get": {
                "tags": [
                    "ATV Images"
                ],
                "summary": "Get all images for a specific ATV",
                "operationId": "4f8d28e2535be53951bc8a3e59b9abdb",
                "parameters": [
                    {
                        "name": "atv_id",
                        "in": "path",
                        "description": "ATV ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by type (image, video)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "image"
                        }
                    },
                    {
                        "name": "active_only",
                        "in": "query",
                        "description": "Show only active images",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ATV images retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AtvImage"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ATV not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ATV not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ATV Images"
                ],
                "summary": "Add an image/video to ATV gallery",
                "operationId": "8e628661ba9c69754f8af1a23d739a65",
                "parameters": [
                    {
                        "name": "atv_id",
                        "in": "path",
                        "description": "ATV ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "url",
                                    "type"
                                ],
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "image",
                                            "video"
                                        ],
                                        "example": "image"
                                    },
                                    "alt_text": {
                                        "type": "string",
                                        "example": "ATV front view"
                                    },
                                    "sort_order": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "is_primary": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Image added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Image added successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AtvImage"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ATV not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ATV not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/atvs/{atv_id}/images/{id}/set-primary": {
            "post": {
                "tags": [
                    "ATV Images"
                ],
                "summary": "Set an image as primary for ATV",
                "operationId": "20f84b7e00603279f5480862d55bd9a0",
                "parameters": [
                    {
                        "name": "atv_id",
                        "in": "path",
                        "description": "ATV ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Image ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Primary image set successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Primary image set successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ATV or image not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ATV or image not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AtvImage": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "atv_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "url": {
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "image",
                            "video"
                        ],
                        "example": "image"
                    },
                    "alt_text": {
                        "type": "string",
                        "example": "ATV front view"
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_primary": {
                        "type": "boolean",
                        "example": false
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}